move files into subdirectories
