+++ src/app/Dashboard/Charts/jfr/JFRMetricsChartController.tsx
