+++ src/app/Dashboard/Dashboard.tsx
@@ -56,10 +56,10 @@ import { Observable, of } from 'rxjs';
 import { AddCard } from './AddCard';
 import { AutomatedAnalysisCardDescriptor } from './AutomatedAnalysis/AutomatedAnalysisCard';
 import { ChartContext } from './Charts/ChartContext';
-import { JFRMetricsChartCardDescriptor } from './Charts/JFRMetricsChartCard';
-import { JFRMetricsChartController } from './Charts/JFRMetricsChartController';
-import { MBeanMetricsChartCardDescriptor } from './Charts/MBeanMetricsChartCard';
-import { MBeanMetricsChartController } from './Charts/MBeanMetricsChartController';
+import { JFRMetricsChartCardDescriptor } from './Charts/jfr/JFRMetricsChartCard';
+import { JFRMetricsChartController } from './Charts/jfr/JFRMetricsChartController';
+import { MBeanMetricsChartCardDescriptor } from './Charts/mbean/MBeanMetricsChartCard';
+import { MBeanMetricsChartController } from './Charts/mbean/MBeanMetricsChartController';
 import { DashboardCard } from './DashboardCard';
 import { DashboardCardActionMenu } from './DashboardCardActionMenu';
 import { QuickStartsCardDescriptor } from './Quickstart/QuickStartsCard';
